feat-02: add structure of the page
