+++ src/components/footer/Footer.jsx
@@ -1,0 +1,12 @@
+import './Footer.css'
+
+function Footer() {
+
+    return (
+        <div>
+            Footer
+        </div>
+    )
+}
+
+export default Footer
