+++ src/components/header/Header.jsx
@@ -1,0 +1,12 @@
+import './Header.css'
+
+function Header() {
+
+    return (
+        <div>
+            Header
+        </div>
+    )
+}
+
+export default Header
